Update sse_mathfun.h
I am not an expert in this subject area, so please excuse me if this pull request is a mistake. However, upon reviewing the code, the comment claims to calculate:

```
if( x < SQRTHF ) {
    e -= 1;
    x = x + x - 1.0;
} else { x = x - 1.0; }
```
And the above code would match the intention from the logf.c code in the cephes library.

However, it appears to actually calculate:
```
if( x < SQRTHF ) {
    e -= 1;
    x = (x-1.0) + (x-1.0);
} else { x = x - 1.0; }
```
Because `1` is subtracted from `x` early in the code.

Again, I am not an expert in this kind of maths & SSE so I apologize if this is due to my misunderstanding, but I thought I would bring it up regardless.